+++ src/assets/util.ts
