+++ src/views/TeamRecordsPage/TeamRecordsPage.tsx
@@ -2,7 +2,7 @@ import { useState, useEffect } from 'react';
 import { Box } from '@mui/material';
 import { getTeamRecords } from '../../services/app.service';
 import Subheader from '../../components/Subheader/Subheader';
-import { callAPI } from '../../assets/helperFunctions';
+import { callAPI } from '../../assets/util';
 import { RecordData } from '../../types';
 
 const TeamRecordsPage = () => {
